+++ src/components/_private/Flyout/FlyoutControlled.tsx
@@ -35,7 +35,7 @@ const FlyoutRoot = (props: T.ControlledProps & T.DefaultProps) => {
 		disableHideAnimation,
 		disableContentHover,
 		disableCloseOnOutsideClick,
-		contentGap,
+		contentGap = 2,
 		contentClassName,
 		contentAttributes,
 		position: passedPosition,
@@ -90,6 +90,7 @@ const FlyoutRoot = (props: T.ControlledProps & T.DefaultProps) => {
 		defaultActive: resolvedActive,
 		container: containerRef?.current,
 		forcePosition,
+		contentGap,
 	});
 	const { status, updatePosition, render, hide, remove, show } = flyout;
 	// Don't create dismissible queue for hover flyout because they close all together on mouseout