+++ src/components/_private/Flyout/utilities/getPositionFallbacks.ts
@@ -1,0 +1,44 @@
+import type * as T from "../Flyout.types";
+
+// All available positions for each side
+const positions: Record<T.Side, T.Position[]> = {
+	top: ["top-start", "top-end", "top"],
+	bottom: ["bottom-start", "bottom-end", "bottom"],
+	start: ["start-top", "start-bottom", "start"],
+	end: ["end-top", "end-bottom", "end"],
+};
+
+// Order of sides to try depending on the starting side
+const fallbackOrder: Record<T.Side, T.Side[]> = {
+	top: ["bottom", "start", "end"],
+	bottom: ["top", "end", "start"],
+	start: ["end", "top", "bottom"],
+	end: ["start", "bottom", "top"],
+};
+
+// Get an order of positions to try to fit flyout on the screen based on its starting position
+const getPositionFallbacks = (position: T.Position) => {
+	const result: T.Position[] = [];
+	const chunks = position.split("-") as [T.Side, string];
+	const [firstChunk] = chunks;
+	const passedPositionOrder = positions[firstChunk];
+	const startingFallbackIndex = passedPositionOrder.indexOf(position);
+	const fallbackIndexOrder = [startingFallbackIndex];
+
+	passedPositionOrder.forEach((_, index) => {
+		if (index === startingFallbackIndex) return;
+		fallbackIndexOrder.push(index);
+	});
+
+	[firstChunk, ...fallbackOrder[firstChunk]].forEach((fallbackSide) => {
+		const fallbackOrder = positions[fallbackSide];
+
+		fallbackIndexOrder.forEach((index) => {
+			result.push(fallbackOrder[index]);
+		});
+	});
+
+	return result;
+};
+
+export default getPositionFallbacks;